website_example_module/lambda_code/requirements.txt
boto3 
pyopenssl